backend/requirements.txt
flask
firebase-admin
google-cloud-aiplatform
python-dotenv
flask-cors
google-cloud-discoveryengine
google-cloud-storage
